Move `CalculateMaxNodesForShoot` to `pkg/gardenlet/operation/botanist/clusterautoscaler.go`
