+++ pkg/utils/gardener/shoot_test.go
@@ -1362,99 +1362,4 @@ var _ = Describe("Shoot", func() {
 		Entry("dual-stack shoot (ipv4 preferred)", &gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{Networking: &gardencorev1beta1.Networking{IPFamilies: []gardencorev1beta1.IPFamily{gardencorev1beta1.IPFamilyIPv4, gardencorev1beta1.IPFamilyIPv6}}}}, "dual-stack"),
 		Entry("dual-stack shoot (ipv6 preferred)", &gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{Networking: &gardencorev1beta1.Networking{IPFamilies: []gardencorev1beta1.IPFamily{gardencorev1beta1.IPFamilyIPv6, gardencorev1beta1.IPFamilyIPv4}}}}, "dual-stack"),
 	)
-
-	DescribeTable("#CalculateMaxNodesForShoot",
-		func(shoot *gardencorev1beta1.Shoot, expectedResult *int64) {
-			maxNode, err := CalculateMaxNodesForShoot(shoot)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(maxNode).To(Equal(expectedResult))
-		},
-
-		Entry(
-			"no network",
-			&gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{
-				Kubernetes: gardencorev1beta1.Kubernetes{
-					KubeControllerManager: &gardencorev1beta1.KubeControllerManagerConfig{
-						NodeCIDRMaskSize: ptr.To[int32](24),
-					},
-				},
-			}},
-			nil,
-		),
-		Entry(
-			"Pods network only",
-			&gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{
-				Kubernetes: gardencorev1beta1.Kubernetes{
-					KubeControllerManager: &gardencorev1beta1.KubeControllerManagerConfig{
-						NodeCIDRMaskSize: ptr.To[int32](24),
-					},
-				},
-				Networking: &gardencorev1beta1.Networking{
-					Pods: ptr.To("100.64.0.0/12"),
-				},
-			}},
-			ptr.To[int64](4096),
-		),
-		Entry(
-			"Default Pods network",
-			&gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{
-				Kubernetes: gardencorev1beta1.Kubernetes{
-					KubeControllerManager: &gardencorev1beta1.KubeControllerManagerConfig{
-						NodeCIDRMaskSize: ptr.To[int32](24),
-					},
-				},
-				Networking: &gardencorev1beta1.Networking{
-					Pods:  ptr.To("100.64.0.0/11"),
-					Nodes: ptr.To("10.250.0.0/16"),
-				},
-			}},
-			ptr.To[int64](8192),
-		),
-		Entry(
-			"Pods network is restriction",
-			&gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{
-				Kubernetes: gardencorev1beta1.Kubernetes{
-					KubeControllerManager: &gardencorev1beta1.KubeControllerManagerConfig{
-						NodeCIDRMaskSize: ptr.To[int32](24),
-					},
-				},
-				Networking: &gardencorev1beta1.Networking{
-					Pods:  ptr.To("100.64.0.0/12"),
-					Nodes: ptr.To("10.250.0.0/16"),
-				},
-			}},
-			ptr.To[int64](4096),
-		),
-		Entry(
-			"Nodes network is restriction",
-			&gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{
-				Kubernetes: gardencorev1beta1.Kubernetes{
-					KubeControllerManager: &gardencorev1beta1.KubeControllerManagerConfig{
-						NodeCIDRMaskSize: ptr.To[int32](24),
-					},
-				},
-				Networking: &gardencorev1beta1.Networking{
-					Pods:  ptr.To("100.64.0.0/11"),
-					Nodes: ptr.To("10.250.0.0/20"),
-				},
-			}},
-			ptr.To[int64](4094),
-		),
-		Entry(
-			"IPv6",
-			&gardencorev1beta1.Shoot{Spec: gardencorev1beta1.ShootSpec{
-				Kubernetes: gardencorev1beta1.Kubernetes{
-					KubeControllerManager: &gardencorev1beta1.KubeControllerManagerConfig{
-						NodeCIDRMaskSize: ptr.To[int32](64),
-					},
-				},
-				Networking: &gardencorev1beta1.Networking{
-					Pods:       ptr.To("2001:db8:1::/48"),
-					Nodes:      ptr.To("2001:db8:2::/48"),
-					IPFamilies: []gardencorev1beta1.IPFamily{gardencorev1beta1.IPFamilyIPv6},
-				},
-			}},
-			ptr.To[int64](65536),
-		),
-	)
 })
